feat: disable user

+++ src/modules/users/resolvers/users.resolver.ts
@@ -8,6 +8,8 @@ import { Roles } from 'src/modules/auth/decorators/role.decorator'
 import { Role } from 'src/global/enums/roles.enum'
 import { UserResultDto } from '../dtos/user-result.dto'
 import { UpdateUserInput } from '../inputs/update-user.input'
+import { DisableUserInput } from '../inputs/disable-user.input'
+import { BadRequestException } from '@nestjs/common'
 
 @Resolver()
 @Roles(Role.manager)
@@ -43,8 +45,27 @@ export class UsersResolver {
   }
 
   @Mutation(() => UserResultDto)
-  async deleteUser(@Args('id') id: string) {
+  async deleteUser(@CurrentUser() currentUser: User, @Args('id') id: string) {
+    if (currentUser.id == id) {
+      throw new BadRequestException('Operation cannot be performed!')
+    }
     const user = await this.usersService.delete(id)
+    return user
+  }
+
+  @Mutation(() => UserResultDto)
+  async disableUser(
+    @CurrentUser() currentUser: User,
+    @Args('disableUserData') disableUserInput: DisableUserInput,
+  ) {
+    if (currentUser.id == disableUserInput.id) {
+      throw new BadRequestException('Operation cannot be performed!')
+    }
+
+    const user = await this.usersService.disableUser(
+      disableUserInput.id,
+      disableUserInput.active,
+    )
     return user
   }
 }
